requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi @ file:///C:/b/abs_91u83siphd/croot/certifi_1700501720658/work/certifi
charset-normalizer==2.0.12
colorama==0.4.5
contourpy==1.2.0
cycler==0.12.1
distro==1.8.0
Django==3.2.23
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
fonttools==4.46.0
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-resources==5.4.0
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
multidict==6.0.4
mysql-connector==2.2.9
mysqlclient==2.2.1
numpy==1.26.2
openai==0.28.0
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.27.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
threadpoolctl==3.2.0
tqdm==4.64.1
typing_extensions==4.8.0
urllib3==1.26.18
yarl==1.9.3
zipp==3.6.0
